Adding a domain/range check for funcheb2/compose.
We can't expect the following to work:
 >> f = sin(x)
 >> g = 100*cos(x)
 >> compose(g, f) % i.e., f(g)
as f is only a good approximation to sin(x) on [-1, 1] and the range of
g is (approx) [-100 100]. Evaluating f at these values will just give
garbage.

Hence, for compositions of the type we estimate the range of g by
looking at its .values field, and throwing an error if norm(g.values, inf)
is greater than 1.