Adding support for complex-valued functions in @funcheb/minandmax.
We follow the approach taken in chebfun v4, whereby we find the
min and max of |f|, but return the return the value of f at these
points.

Furthermore, we avoid introducing breakpoints by locating the
min and max of real(f)^2 + imag(f)^2, rather than |f|.

Modified the test functions accordingly.