Adding basic implementations of @chebfun, @fun, and @bndfun classes.
This commit contains (essentially) the minimal files required to
construct a piecewise chebfun on a finite domain (with no blow up).

The splitting technology is contained in @fun/detectedge, and is
copied almost directly from ChebfunV4.

FEVAL, OVERLAP, CHEBPOLYPLOT and a (very) basic PLOT have been
implemented to aid others in further developments.