Correct the factor for barycentric weight when using first kind formula.
funcheb1.bary and funcheb2.bary decide the factor and pass it to funcheb.bary
and this factor is passed to bary1 by the wrapper in funcheb.bary.

Also the barycentric weights have been adjusted so that the first entry and the
last entry correspond to the left and right end points respectively.