Removing everything besides the @funchebX class and trunk methods.
The first commit to the master chebfun repo should only contain
those files related to @funcheb classes. Here I am temporarily
removing them from my fork in order for this to happen. The commit
history will therefore be a bit of a mess, but this shouldn't be
too much of a problem.

Note that we have also removed the statement causing @funcheb to
inherit from @smoothfun, as we have removed that class from the
system for now.

I will shortly be making a branch which contains those files which
have just been deleted.