Adding plus, minus and uminus functionality for bndfuns.
We still need to decide how to deal with checking whether the intervals
match, and how the tolerance for the check is determined by hscale. Perhaps
there should be a private method for bndfuns which takes in the required
information and returns whether they match, as that will be an operation
required for most bndfun methods.