Begin merging funcheb1 and funcheb2 common tests.
We have copied over the bulk of the funcheb2 tests and added loops to run them
for funcheb1 objects as well.  This code is undeniably less readable than it
was previously, but there isn't a good way to deduplicate them otherwise.

Also, it seems that some of the tests do not pass for funcheb1 as-is, likely
due to tolerances.  This is on our list of issues to address next.