Factor out logic common to per-class test() methods.
Currently, both funcheb1 and funcheb2 have methods that duplicate the exact
same switch-to-test-directory-and-loop-through-files logic.  We create a new
function named chebtest() to replace them, which lives in the project root (no
better place for it at the moment).  This function accepts subdirectories of
tests/ as string arguments and runs all the tests that reside in those
directories.  If given no arguments, it runs the tests in all of the tests/
subdirectories.