Break symmetry in tests/funcheb/test_max() to force unique answer.
The previous function turned out to attain its global maximum (maximum modulus
in this case) at more than one point.  funcheb1 and funcheb2 picked different
points.  We replaced it with a function that has a unique maximum to avoid this
issue.

We note that this problem does not plague the test functions used in
test_min().