Merge test_bary().
A subtlety:  this test doesn't actually call funcheb.bary().  It calls
funcheb1.bary() and funcheb2.bary(), which in turn call funcheb.bary().  This
is necessary because only the subclass bary() functions know how to set the
scaling parameter 'p' required for funcheb.bary() to work.

It would be good if we could test these things separately, but this doesn't
seem to be the case.