Modifying feval to use clenshaw's algorithm to evaluate off the real line.
Clenshaw's algorithm is noticably faster for evaluation off the real line
than the barycentric interpolation formula (by approximately a factor of
three!) and also appears to be more accurate than the first form formula
(as advocated by Webb, Gonnet, and Trefethen).

Bary1 is no longer needed, and has therefore been removed from bary. Hence
there is is no need to pass the 'kind' flag, as the only available kind is
2.

Furthermore, bary and clenshaw will now only accept a column vector as an
input. If they are given something else, it will reshape it with x(:) and
throw a warning.

Documentation amended accordingly.