Update chebtech ishappy field in plus, times.
If we don't do this, then we end up with absurdities, e.g., adding a happy
function to an unhappy one makes the result happy.  Tests have been added to
check for this.  Thanks to Georges Klein for pointing this out!

This is probably not the end of the story here, though, as the following merits
consideration.  Suppose we have an unhappy chebtech that is very small, say,
with a vscale of 1e-6.  We add to it a chebtech which is constant (and
therefore happy) with a value of 2^300.  The unhappy summand is just noise in
the sum.  Should the sum not therefore be happy?