Modify tests to be compatible with both old and new RNG seeding syntax.
This is accomplished by adding a new seedRNG() function which checks the MATLAB
version and acts accordingly.  Currently this function lives in the project
root because there isn't really a better place for it, but that can be changed
freely later.

We also no longer guarantee that the RNG state will be restored on the exit
from a test.  The reasons for this are that it's extra work for us to maintain
the code, would require the introduction of more of these compatibility
functions, and it likely won't matter much to anyone, since it's highly
improbably that someone will call chebtest() in the middle of a statistical
simulation...

This commit is an update to the changes made previously in commit 85d19f07.