Modifications to PROLONG, ALIAS, and ROOTS as suggested by Townsend.
The major algorithmic change in ROOTS is that
  tailMmax = eps*norm(c, 1);
instead of
  tailMmax = 1e-15*norm(c, 1);
AT suggests that the new approach is just as stable and more accurate.

He also suggests that we should truncate (rather than alias) coeffs
in ROOTS (as aliasing is a value-space property) and that we use an
upper-Hessenberg form for A. He will investigate these further.