Adding experimental way of approximating gradient of a chebfun in construction.
This removes the need for the finte difference approximation, and hence for hscale.

To proceed, we compute the coefficients of the derivative, and approximate the maximum
value using the bound |f(x)| < sum(abs(f.coeffs)). This is very crude, but typically
it gives a result of the correct order when n is small/medium. To correct for large n,
we actually use sum(abs(f.coeffs)/log(length(f.coeffs)), which can be shown to give a
good approximation for the finite difference gradient approximation to sin(n*x).