Some changes to barymat(), courtesy of Mohsin Javed.
The most notable change is that the input validation check at the start of the
file has been changed to
  if ( size(x, 2) > 1 || size(y, 2) > 1 )
from
  if ( min(size(x)) > 1 || min(size(y)) > 1 )

According to the error message triggered if this condition is violated, the
check is supposed to ensure that the inputs are column vectors.  The previous
check does not actually do this.

All tests pass, though we don't have a test for barymat().  In particular the
test for chebtech.qr(), which uses barymat(), passes.