Changes to how mappings are dealt with.
Introducing an abstract, static method for FUN, createMap(), which is to map
[-1, 1] to the domain of the FUN object (which could be a BNDFUN or an
UNBDNFUN). For BNDFUN objects, this is the previous linear.m method, which
has now been renamed to createMap.

As a part of this, removing linear.m from the FUN level, which should not be
a function for that class.