A simpler approach to simplify (from Nick Hale).
simplify() will now just zero out small coefficients (where "small" is either
the default eps preference or a user-specified tolerance) and trim the tail as
necessary.  Calling happinessCheck() inside simplify() and basing the
simplification on that as we did previously doesn't make much sense because
simplify() cannot grow the length of the representation.