Some fixes related to new implementation of SIMPLIFY() and QR().
Some tests were throwing warnings about simplifying unhappy
CHEBTECH objects, this was because IMAG(F) was returning an
unhappy CHEBTECH even when F was real. I fixed this by setting
f.ishappy = 1 after calling f.make, and perhaps this needs to
be done elsewhere also.

QR was failing in the degenerate case because of the way we
were trying to enforce positive diagonal elements in R. This
has now been fixed.

Some tidying of hep text in SIMPLIFY(), in particular stating
that a warning will be thrown if an unhappy CHEBTECH is passed.
Codes calling SIMPLIFY() should be made to check for happiness
before calling (or perhaps we allow a 'no warning' flag?)