Removing the checkDomain method of FUN.
Binary methods at the FUN/BNDFUN level will assume that both arguments actually
live on the same domain. No warning is thrown if not, but the results will be
gibberish. This is now documented in the both the classdef files of FUN and
BNDFUN, as well as in all binary methods for these classes.