Revise error message in chebtech.compose().
This should help this make more sense if this error is triggered, e.g., in the
context of a call to bndfun.compose().  All of this is very confusing to me.
We should make sure to bring up the question of how to handle errors that
travel between layers like this with Mr. Cherrie when he comes by.