Tidying up the initial version of cumsum@singfun. Some of the functions/routines inherited from Chebfun v4 have been commented out, including makelog, jacsum, and jac2cheb, since they may not be useful in the future. But they will stay there for time being for they may revive in the future.
WARNING: The current version of cumsum@singfun can cover limited cases. The case
 that the antiderivative has logarithm (including atanh) and the case that the
integrand has singularities at both ends are not supported.

TODO: Improvement on the algorithm to handle the case with singularities at
both the end points. Improvement on the whole singfun class to handle
logarithmic singularity, otherwise the singfun class is not closed. Here, by
'closed' we mean that the result of an operator on an arbitrary singfun (unary
operator) or any two arbitrary singfuns (binary operator) is representable by
a singfun object.