Modify chebfun.feval() to work for general N-dimensional inputs.
There are two subtleties here:

1. feval() internally turns its input to be a column vector.  This is
fine, but if the input was multidimensional with dimensions d1 x ... x
dm, then the output cannot be turned into an array of the right
dimensions with a single call to reshape.  Looking at the output array,
it becomes clear that the result needs to be transposed first but in
such a way that the if the matrix is broken into (d1 d2) x 1 blocks,
those blocks are kept intact.  We accomplish this by using mat2cell() to
turn the matrix into a cell array of these blocks and then transpose
that cell array.

2. If the chebfun being evaluated is transposed, the output is also
transposed.  If the output is two-dimensional or less, this makes sense,
but if it has higher dimensions, it is unclear what the right way to
interpret this should be.  We've defined the "transpose" of a
multidimensional array to mean the same array but with its first two
dimensions swapped.

The test for @chebfun/feval.m has been updated accordingly.