Make chebfun.chebpolyplot() work with array-valued chebfuns with breakpoints.
We plot one curve for each fun of each column of the chebfun.  If the chebfun
has lots of columns, lots of funs, or, especially, lots of both, the plot is a
mess, but I'm not sure there's a better way to do this.

I have a feeling that the changes in this code were made necessary due to
breakage when we changed how information is stored in chebfun, but I haven't
investigated this.