Fix bug in chebtech2.compose().
chebtech2.compose() crashes when applying a binary operator to a pair of
array-valued chebtechs if the resampling and extrapolate flags are set due to a
missing ':' in a line accessing one variable.  This commit fixes this issue.
The test for chebtech.compose() has been modified to try things with the
extrapolate flag set to catch problems like this in the future.