Tweak tolerances to fix tests which broke due to changes in previous commit.
While we were at it, we fixed most of the tests for the chebtech classes which
did not set tolerances properly based on vscales and epslevels.  There are a
few tests for which we were not sure if conversion was necessary, so we left
them alone.  It seems that there are lots of tests in fun and bndfun that do
not do this correctly either.