Fix bug in call to chebfun.merge() during construction.
This transposition should not be here.  Weirdly enough, it doesn't
always cause a problem, but sometimes, if there are multiple breakpoints
involved, an error occurs.  The example function we used to draw out
this error is

f = chebfun(@(x) sign(x - 0.1).*abs(x + 0.2).*sin(3*x), 'splitting', 'on');

A test involving the construction of this function has been added to
check for this problem in the future.