Fix bug in call to chebfun.merge() during construction.
Cherry-pick 1fc3688d into this branch so we can do a reasonable test of
the trig functions.  Even with manually-introduced breakpoints, the
subsequent call to compose() with the trig function can result in new
breakpoints that have to be merged, and without this fix, the call to
merge() will result in a crash.