Fix bug in call to chebfun.merge() during construction.
OK, I think I have it right this time.  The issue which commit 1fc3688d
was made to correct arises because of the change in behavior of MATLAB's
setdiff() in R2013a.  This explains why the change from that commit
breaks things in R2013a, necessitating the reversion in 416d265b.  The
point is that the index input to merge() _must_ be a row vector.
Turning it into a column before transposing with "(:)" should allow us
to do what we want.