Fix bugs in test/fun/test_minandmax.m.
This test started failing on my personal machine after the changes to
chebtech.simplify() introduced in 2b1fb833, but these changes weren't
the real culprit.  There were two bugs here:

1. The real test 6 was being "shadowed" by test 7 due to test 7
mistakenly storing its result in position 6 of the pass vector.

2. Test 7 computes the extrema of a complex function twice:  once with
that function in a fun of its own and once with that function as part of
an array-valued fun.  The extrema of this particular function are
nonunique, occurring at several points in the domain, and moreover, it
has _two_ maximum extreme values (i.e., values with maximum modulus on
the domain): +1 and -1.  Test 7 was checking to see that the _values_
obtained were the same in both cases, while the best you can hope for is
that the _moduli of the values_ coincide due to the way extrema of
complex-valued functions are defined.  The test for chebtech.minandmax()
does have this problem.