Update vscale in chebtech.diff() when taking differences across columns.
This should probably technically be done on development, and then
cherry-picked into this branch, but I don't think it will be an issue
elsewhere, so I've just placed it here.

chebtech.diff() was not updating vscale after doing a diff across
columns.  The reason we didn't detect this earlier is probably because
most of the chebtech tests don't use vscale and epslevel to define their
tolerances.  I've updated the test for chebtech.diff() to verify at
least that the vscale field has been correctly resized.

Ostenisbly, the epslevel needs to be updated here as well, but I'm  not
sure how that should be handled, so I've marked it with a TODO for now.