Fix chebfun.feval() for evalution at vector which contains repeated breakpoints.
With the code prior to this commit, doing

  >> f = chebfun(@(x) [sin(x) cos(x)])
  >> feval(f, [-1 ; -1])

results in a dimension mismatch error because the changed line attempts to
assign the two-column vector with the values of the function at the impulses to
the 2-by-2 slice of fx determined by the index variable.  This problem does not
occur for scalar-valued chebfuns because of how MATLAB handles assigning
scalars to slices of matrices.  This commit fixes this by ensuring that the
vector of impulse values is vertically replicated as many times as needed prior
to assignment.

A test has been added to ensure that this issue does not arise again in the
future.