Adding @CHEBFUN/CHEBPOLY() and @FUN/CHEBPOLY() methods.
These aare designed for returning the Chebyshev coefficients
of a possibly piecewise of singular function. i.e., they do
not rely on the fact that the underlying representation is a
CHEBTECH, but rather compute the coefficients by computing
the relevent inner products (if required).

TODO: Implement this as SINGFUN level (which should be as
simple as sum(f.*T_k./sqrt(1-x.^2)), k = 0, ..., N.