Redesign of @CHEBFUN/ABS().
ABS() simply calls ADDBREAKSATROOTS() and then farms
the rest of the work off to the FUNS (and in turn the
underlying tech).

SIGN() isn't quite the same, as we don't want to waste
time restricting when we only want a logical CHEBFUN
as an output. However, it could probably still use the
new ADDBREAKSATROOTS() method.

I've added a test for complex F in ABS, which fails
becuse the constructor does not detect a plateau in
the coeffs...