Fix addition of array-valued "scalars" to array-valued chebtechs.
9efc5160 and 085d0dba were made with two ideas in mind:  (1) scalar
expansion for chebfuns so that we can do things like, e.g., f + [1 2] to
get a chebfun for @(x) [(1 + f(x)) (2 + f(x))] and (2) eliminating the
need for b39180e0 (on feature-chebfun-calculus) by allowing the addition
of array-valued "scalars" to array-valued chebfuns, e.g, if f represents
@(x) [sin(x) cos(x)], then f + [1 2] represents @(x) [(1 + sin(x)) (2 +
cos(x))].

(2) was already supported in chebtech, and the only thing necessary to
get it to work with chebfun was to change how the impulses were updated.
(1) required a change to chebtech.  The change to chebtech to support
(1) ended up breaking support for (2) in chebtech.  This commit fixes
that and adds tests for chebtech.plus() and chebfun.plus() that exercise
this behavior so that if we break it again, we will know.