Subsitute zero for empty third argument in chebfun.residue().
This is a convenient shortcut which allows you to write

  >> [B, A] = residue([1 1], [1 -1], chebfun())

instead of the slightly more verbose

  >> [B, A] = residue([1 1], [1 -1], chebfun(@(x) 0*x)).

It's even in-line with MATLAB's syntax for residue:  residue(r, p, 0) and
residue(r, p, []) do the same thing, at least as far as I can see!

This improvement was suggested by Georges Klein.