Use FFT only in chebtech.prolong().
Previously, we would switch between the FFT and bary() depending on the size of
the prolongation and the length of the chebtech involved.  In the last commit,
we swtiched to using clenshaw() instead of bary() to be more consistent with
the rest of chebtech but did not update the swtich-over criterion.

Nick Hale ran some performance tests and found that there is basically no
advantage to using clenshaw() over the FFT, while bary() enjoys a marginal
advantage over the FFT in the regime where nIn is small and nOut is large.
Since we're no longer using bary(), and since there's nothing to be gained by
switching between the FFT and clenshaw(), we have decided to keep the code
simple and just use the FFT 100% of the time.