Add function for zeroing small breakpoint values.
Say we do

  >> f = chebfun(@(x) exp(x));
  >> g = chebfun(@(x) (exp(0.5) - exp(-0.5))*(x + 0.5) + exp(-0.5));
  >> h = sign(f - g);
  >> [h.domain.' h.impulses]
  ans =
    -1.000000000000000   1.000000000000000
    -0.500000000000001                   0
     0.500000000000000                   0
     1.000000000000000   1.000000000000000

This is fine.  (Note that g is the polynomial of degree 1 which interpolates
exp() at +/- 0.5.)  On the other hand, if we specify +/- 0.5 as breakpoints
when constructing f, we get

  >> f = chebfun(@(x) exp(x), [-1 -0.5 0 0.5 1]);
  >> h = sign(f - g);
  >> [h.domain.' h.impulses]
  ans =
    -1.000000000000000   1.000000000000000
    -0.500000000000000  -1.000000000000000
                     0  -1.000000000000000
     0.500000000000000                   0
     1.000000000000000   1.000000000000000

which is also fine, except that the impulse value for h at -0.5 is now -1
instead of 0.  This happens because, unlike in the previous example, we didn't
use addBreaksAtRoots() to introduce these breakpoints, so they never get set to
zero explicitly.  As far as the quality of h as a function representation of
sign(f - g) is concerned, this is fine, but functions like eq() rely on the
impulse values at +/- 0.5 being exactly zero.

This commit solves this problem by introducing a function named
thresholdBreakpointValues(), which sets small breakpoint values to zero.  By
calling this function at the end of minus() (or, really, plus()), we can get
the behavior that we need:

  >>  f = chebfun(@(x) exp(x), [-1 -0.5 0 0.5 1]);
  >>  g = chebfun(@(x) (exp(0.5) - exp(-0.5))*(x + 0.5) + exp(-0.5));
  >>  h = sign(f - g);
  >>  [h.domain.' h.impulses]
  ans =
    -1.000000000000000   1.000000000000000
    -0.500000000000000                   0
                     0  -1.000000000000000
     0.500000000000000                   0
     1.000000000000000   1.000000000000000