Use right- and left-hand limits at endpoints when merging funs.
Presently, merge() fails to remove redundant breakpoints in chebfuns like the
following:

  >> f = chebfun({@(x) 0*x + 1, @(x) 0*x + 1, @(x) 0*x}, [-1 -0.5 0 1]);
  >> g = merge(f)
  g =
     chebfun column (3 smooth pieces)
         interval       length   endpoint values
  [      -1,    -0.5]        1         1        1
  [    -0.5,       0]        1         1        1
  [       0,       1]        1         0        0
  Epslevel = 6.492624e-16.  Vscale = 1.  Total length = 3.

g still has three breakpoints, even though the breakpoint at -0.5 is clearly
unnecessary.  The reason is that the the impulse value of f at 0 is 0.5, so if
we try to build a merged fun for [-1, 0] based on chebtech2 (or any tech which
samples at endpoints), it will fail.

The fix is to construct the fun using a dummy chebfun which replaces the
impulse values at the endpoints of the new interval with the appropriate right-
and left-hand limits.  With this patch, we get

  >> g = merge(f)
  g =
     chebfun column (2 smooth pieces)
         interval       length   endpoint values
  [      -1,       0]        1         1        1
  [       0,       1]        1         0        0
  Epslevel = 6.492624e-16.  Vscale = 1.  Total length = 2.

as desired.