Fix bug in chebfun.qr() for chebfuns with one column and multiple funs.
The logic that seeds the recurrence relation used to generate the
Legendre-Chebyshev-Vandermonde matrix in qrSimple() implicitly assumes that
we're working with at least two columns.  This was causing problems when
calling qr() on a chebfun with one column and multiple funs.  (Chebfuns with
one fun were fine, as they were handled by the chebtech qr() code.) This commit
handles the single-column case separately (which we probably wanted to do
anyway, since it's trivial).