Fix new preferences for chebtech.compose().
Really, this is just a matter of making sure default preferences get filled in
at the start of compose(), but we've gone a little further and also changed the
static function chebpref.mergePrefs() can now accept chebpref arguments instead
of just plain structs.  The behavior in this case is to discard all but the
techPrefs substructure, i.e., if p is a chebpref, chebpref.mergePrefs(p, q) is
the same thing as chebpref.mergePrefs(p.techPrefs, q) and similarly if q is a
chebpref.  This defines a way of "flattening" a chebpref into a structure that
can be interpreted by techs.