Fix bug in chebfun.subsref() for evaluating on non-column inputs.
The code supporting, e.g., the syntax f(x, colIndex) was returning just the
columns colIndex of the result.  This works fine if x is a column vector, but
it fails if x is a row vector or array.  This commit fixes this issue by
properly calculating which columns need to be selected based on the size of the
input.

The test for subsref() has been updated accordingly.