Rename *Samples chebtech preferences to *Points.
As NH points out, calling these preferences "maxSamples", etc. is a little
disingenuous, as if you're using a non-nested refinement function (which is
_always_ the case for chebtech1 at the moment), you will end up sampling
(evaluating) the function many more times than this number would suggest.
These preferences really describe limits on size of the Chebyshev grid (number
of points) used in the final representation of the function, so we have renamed
them accordingly.

This commit also fixes a few stray references to "numSamples" tech preferences
in chebfun that we must have missed earlier.