Settle length vs. degree issue in chebpref preferences.
chebpref preferences named *Length now represent quantities that are returned
by length().  For chebtech, the length is the total number of points, or one
plus the degree of the interpolating polynomial.  Previously, there was some
inconsistency here, as the higher-level preferences were cast in terms of
polynomial degree while the lower-level ones were in terms of number of points,
resulting in the need to remember to add one to these values here and there
inside chebfun.  This commit fixes this inconsistency.