Port ratinterp() code over from Chebfun v4.
This code at least give the appearance that it works but hasn't been
tested.  Line count is way up from the v4 code, but that's because the
spacing used in the v4 code is aimed at keeping the implementation
compact at the expense of making the code's structure easy to identify.
This code is spaced according to the v5 coding standards, and the body
is broken down into subfunctions, which should make it a little easier
to read.

The main difference between this code and the v4 ratinterp() routine is
in how the domain argument is handled.  In v4, domains were separate
classes; in current v5, we have to make do with row vectors.  This has
unfortunately made the input parsing a bit more complex, but it appears
to work, at least for now.