Fix to chebfun.legpoly() for row chebfuns.
size(f, 2) returns Inf for a row chebfun, which causes this line to break.  No
problem:  what we really wanted was the number of components, so we replace
this with the min(size(f)) idiom that we've been using for this.

We really need to write a numComponents() or similar function somewhere...