Fix breakage of logical functions after merge with development.
Mostly this is just a lot of updates to files for the new preferences
stuff.  Also added some empty any(), and(), and or() functions for
singfun, which must have them now, as they are required by onefun.  We
will need to develop proper implementations for these in the future
(should be easy).