Add support for constructing a chebfun from Chebyshev coefficients.
This is needed by ratinterp().  It works, but it is a little hack-ish.
Then again, so is support for construction from coefficients in
chebtech, so maybe this isn't so bad.

The reason this works is that chebtech interprets an 'op' input of {[],
c} as a list of Chebyshev coefficients, and this is respected by fun,
bndfun, etc.  We then have to wrap it in one more cell to let the
chebfun constructor know that we want to construct on a single interval
with an 'op' of {[], c}, not on two intervals with 'op's of [] and c,
respectively.