Import padeapprox.m code from Chebfun v4.
No real changes; just some reformatting from the ultra-compact version
prepared for publication and some editing for the documentation to put
it into a style consistent with the rest of the project.

We've also added a trivial one-case test file to make sure that things
are working correctly.  It probably isn't necessary to to do much more
than this.  padeapprox() was not tested at all in Chebfun v4, likely
because it's not _really_ a Chebfun code.  The test for chebpade(), on
the other hand, will be more extensive.