Port chebfun.cf() over from Chebfun v4.
The test isn't as thorough as we would like, but it hits the high
points.  The code could also probably use some further commenting, but
anything much more detailed than what is already written will require a
lot of time and effort to figure out the details of the algorithm.
Fortunately, there are already several good papers by LNT and Gutknecht
on this method, and because of that, it may be a good idea to just leave
it as-is for the time being and focus on higher-priority targets.