Bring inv2() online.
There were two problems preventing this from working:

1.  We were throwing an error when the number of Newton iterations
exceeded 100 instead of just exiting the loop and moving on to the next
point.  Setting the iteration limit to 10 and replacing the error with a
break gives results that look fine except for the value at the right
endpoint (see next paragraph).

2.  At some point in the construction of the chebfun for the inverse of
f, we call getValuesAtBreakpoints([a b]), where a and b are the left and
right endpoints of the domain of that chebfun.  Unfortunately, the
Newton iteration in fInverse() derives its initial guesses from the
solution for the previous entry in the vector x of sampling points.  If
the input vector is not sufficiently dense in the construction domain,
the guesses can be too far off from the actual solutions, causing the
Newton iteration to lock onto wrong solutions.

To fix this, we use the fInverse from inv() to do the computation if the
vector of sample points is not long enough.  Since fInverse is only ever
called during the construction process (and hence not with arbitrary
vectors of sampling points), this will hopefully be OK, at least if we
are using an underlying tech like chebtech, which is based on sampling
on finer and finer grids.

These issues will be documented inline in a future commit once we've
refactored inv() and inv2().