Refactor inv() and inv2().
inv2.m was basically just a copy of inv.m with just the body of the
fInverse() routine changed.  All code has now been moved into inv.m, and
you can select between the different options for fInverse via a new
'algorithm' flag.  The body of inv2() has been replaced with a call to
inv() with this flag set appropriately.

All-in-all, this results in about 50-60 fewer lines of code than the
previous configuration.