Fix "chebpts1" and "chebpts2" flags.
These flags were trying to alter the choice of Chebyshev grid by setting
the "tech" preference to "chebtech1" and "chebtech2".  This worked in
the pre-chebpref world (prior to commit f165262c); however, since then,
the correct way to select the grid type has been to set the tech to
"chebtech" and use the "gridType" chebtech preference to switch between
them.