Create new chebfun.getRootsForBreaks() function.
This function computes the roots of a chebfun and then cleans them up in such a
way that they are suitable for use as breakpoints in a chebfun with the same
domain.  We need this for computing, e.g., log1p(f), which requires breakpoints
to be introduced into f at the roots of 1 + f.

Suggestions for better a better name are welcome.