Add breaks at correct roots in chebfun.log1p().
f = addBreaksAtRoots(f + 1) replaces f with a copy of _f + 1_ with breakpoints
at its roots, not f.  It is tempting to fix this by just subtracting 1 from the
result; however, this defeats the purpose of the function log1p in the first
place:  the argument is supposed to be small, and therefore doing (f + 1) - 1
will cause you to lose digits.  Instead, we use the new getRootsForBreaks() and
addBreaks() functions to get the job done.