Fix bug in plotting of complex-valued array-valued chebfuns.
If the chebfun is complex, dataNew.xLine gets replaced by
real(dataNew.yLine), so it is not one-dimensional if the chebfun is
array-valued, causing the padding with a single NaN value to fail.  The
solution is to pad with a NaN vector instead.

This resolves #120 on the GitHub issues tracker.