Add primitive automated test for plot().
This addresses #122 on the GitHub issues tracker.  We can't actually check that
the plots are accurate without a human in the loop (or some much more
sophisticated processing), but we can at the very least check that none of the
syntaxes triggers a crash.

Currently, plotting scalar chebfuns against array-valued ones (tests 20 and 21)
causes a crash, while plotting scalar chebfuns against quasimatrices (tests 18
and 19) does not.  This is a bug which demands fixing.