Remove support for fix() and floor() for chebfuns with singular funs.
These functions are useless if a chebfun has a fun which is a singfun.
The reason is that these functions require computing the roots of f -
k, where k is an integer, and singfun is not currently well-equipped
to deal with functions of this form (e.g., the usual example of
sqrt(x) - 1).