Fix bug in chebtech constructor involving minPoints and maxPoints.
If, say, pref.minPoints and pref.maxPoints are both set to 8, the refine()
functions in chebtech[12] will exit without ever sampling any values because
minPoints gets rounded up to 9 (to be of the form 2^n + 1), which exceeds
maxPoints.  This causes the loop in chebtech.populate() to terminate on its
first run (giveUp is true), causing the function to crash later on down the
line because epslevel, coeffs, etc. were never set.

This patch (from Nick Hale) fixes this issue by ensuring that at least one set
of samples is always taken.  If minPoints rounds up to be greater than
maxPoints, we just take maxPoints samples.

The chebtech constructor tests have been modified to check that the
construction procedure does not crash under these circumstances with the fix in
place.