Clean some of the v4 cruft out of lebesgue().
Replaced the setting of minPoints and maxPoints with setting of exactLength.
This is more transparent for the reason explained by the inline comment.  It
also allows us to get rid of the code which disables the convergence failure
warnings, as constructions is no longer adaptive.

These changes were suggested by Nick Hale.