Properly handle vscales of array-valued chebtechs in chebtech.qr().
This closes #137.  Changes made in commit 15f80324 replaced these lines with
ones that "scalarize" the vscale.  chebfun.display() was expecting to receive a
vector vscale, and this is what caused the crash noted in the bug report.  This
commit undoes those changes.  We note that the other similar changes made in
15f80324 were all undone elsewhere.