Fix crash in ratinterp() when using roots of unity with constant denominator.
If the length of the denominator coefficient vector (b) is 1, then MATLAB's
conventions for ifft() mean that the call to ifft() in the line changed will
produce a row vector as its output, causing the code to crash with a dimension
mismatch error.  The change in this commit fixes this by forcing the vector
returned by this call to be a column.

This sounds useless, as it is unlikely that anyone would want to use ratinterp
with a degree-zero denominator over just calling the chebfun constructor;
however if robustness is on, and the if the poles of the function being
approximated are far enough from the interpolation points), it may happen that
the robustness features of ratinterp() decide to set the denominator degree to
zero anyway.  A test has been added which deliberately triggers this behavior
to help prevent this problem from resurfacing later on.

This issue does not appear to affect interpolation in other types of nodes.

NB:  This issue is also present in ratinterp() in Chebfun v4.