Check for deltafuns in the isSimple test in chebfun.qr().
If we don't do this, we'll pass into the case where isSimple is true and call
quasi2cheb() to convert the input quasimatrix into an array-valued chebfun.
This conversion is impossible, causing quasi2cheb() to just return the input
quasimatrix.  We then call qr() of the result, taking us back to square one.
This continues until we hit MATLAB's call-stack limit for recursion, causing
things to blow up.

Besides this, we also fix the A.domain --> domain(A) error from ticket #231.
We can actually reach this clause now, since chebfuns based on deltafun are not
"simple".  Of course, since the concept doesn't really make sense, the result
of calling QR on a quasimatrix with a deltafun is still an error, but it's at
least a little less mysterious than what one would get otherwise.