Revert to Chebfun v4 method for determining plot x-limits.
We have been trying to develop a more sophisticated way to approach this, but
we can't seem to get it to work well enough to merit inclusion.  Things will
work for some functions but not others; we can't find something that gives
a reasonably good result universally.  So, we will revert to the old fixed
window style for now and try again later.