Fix plotting test failures introduced by previous commit.
I'm not sure if we want to check for the presence of the xDeltas and
yDeltas fields like I've done here or if we just want to always ensure
that these fields are present.  As the former is simpler and the latter
doesn't really seem to be an improvement, I've gone with the former.