Remove unnecessary properties and code from chebfunpref.
The properties here are no longer necessary:  all preferences are stored
in the prefList struct.  Moreover, it is both more error-prone and not
necessary to initialize outPref field-by-field in the constructor.
These lines were originally removed in 410c94d7 and 2b318699 and appear
to have been accidentally reintroduced in 3f354657.